requirements.txt
bingx>=0.0.101
cryptography>=2.6.1
aiohttp>=3.8
aiodns>=1.1.1
requests>=2.18.4
certifi>=2018.1.18
yarl>=1.7.2
attrs>=17.3.0
multidict>=4.5
frozenlist>=1.1.1
aiosignal>=1.1.2
pycares>=4.0.0
setuptools>=60.9.0
typing-extensions>=4.4.0
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
schedule
